tests/pyproject.toml
[project]
name = "openchoreo-integration-tests"
version = "0.1.0"
description = "Integration tests for OpenChoreo platform"
requires-python = ">=3.11"
dependencies = [
    # Kubernetes
    "kubernetes>=29.0.0",
    
    # HTTP/REST
    "requests>=2.31.0",
    "httpx>=0.27.0",
    
    # Authentication
    "authlib>=1.3.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    
    # Vault/OpenBao
    "hvac>=2.1.0",
    
    # OpenSearch
    "opensearch-py>=2.4.0",
    
    # Prometheus
    "prometheus-api-client>=0.5.4",
    
    # Argo Workflows
    "argo-workflows>=6.5.0",
    
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-timeout>=2.3.0",
    "pytest-html>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "smoke: Quick health check tests",
    "integration: Full integration tests",
    "slow: Tests that take longer to run",
    "infrastructure: Infrastructure layer tests",
    "control_plane: Control plane tests",
    "data_plane: Data plane tests",
    "workflow_plane: Workflow plane tests",
    "observability: Observability plane tests",
    "gitops: GitOps tests",
]
asyncio_mode = "auto"
timeout = 120

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
